Conflicts:
	sefaria/texts.py

+++ sefaria/export.py
@@ -6,7 +6,9 @@ Exports to the directory specified in SEFARIA_DATA_PATH.
 
 import sys
 import os
+import csv
 import simplejson as json
+import re
 from shutil import rmtree
 
 from texts import *
@@ -16,11 +18,11 @@ from database import db
 # To allow these files to be run from command line
 os.environ['DJANGO_SETTINGS_MODULE'] = "settings"
 
-
 lang_codes = {
 	"he": "Hebrew",
 	"en": "English"
 }
+
 
 def make_path(doc, format):
 	"""
@@ -28,21 +30,27 @@ def make_path(doc, format):
 	"""
 	if doc["categories"][0] not in order:
 		doc["categories"].insert(0, "Other")
-	path = "%s/%s/%s/%s/%s/%s.%s" % (SEFARIA_DATA_PATH, 
-										format, 
-										"/".join(doc["categories"]), 
-										doc["title"],
-										lang_codes[doc["language"]],
-										doc["versionTitle"],
-										format)
+	path = "%s/%s/%s/%s/%s/%s.%s" % (SEFARIA_DATA_PATH,
+									 format,
+									 "/".join(doc["categories"]),
+									 doc["title"],
+									 lang_codes[doc["language"]],
+									 remove_illegal_file_chars(doc["versionTitle"]),
+									 format)
 	return path
+
+
+def remove_illegal_file_chars(filename_str):
+	p = re.compile('[/:()<>"|?*]|(\\\)')
+	new_filename = p.sub('', filename_str)
+	return new_filename
 
 
 def make_json(doc):
 	"""
 	Exports doc as JSON (as is).
 	"""
-	return json.dumps(doc, indent=4, ensure_ascii=False) 
+	return json.dumps(doc, indent=4, ensure_ascii=False)
 
 
 def make_text(doc):
@@ -53,11 +61,11 @@ def make_text(doc):
 
 	def flatten(text, sectionNames):
 		if len(sectionNames) == 1:
-			text = [t if t else "" for t in text ]
+			text = [t if t else "" for t in text]
 			return "\n".join(text)
 		flat = ""
 		for i in range(len(text)):
-			section = section_to_daf(i+1) if sectionNames[0] == "Daf" else str(i+1)
+			section = section_to_daf(i + 1) if sectionNames[0] == "Daf" else str(i + 1)
 			flat += "\n\n%s %s\n\n%s" % (sectionNames[0], section, flatten(text[i], sectionNames[1:]))
 		return flat
 
@@ -72,9 +80,9 @@ The name is used as a top level directory and file suffix.
 The function takes a document and returns the text to output.
 """
 export_formats = (
-		('json', make_json),
-		('txt',  make_text),
-	)
+	('json', make_json),
+	('txt', make_text),
+)
 
 
 def clear_exports():
@@ -82,7 +90,28 @@ def clear_exports():
 	Deletes all files from any export directory listed in export_formats.
 	"""
 	for format in export_formats:
-		rmtree(SEFARIA_DATA_PATH + "/" + format[0])
+		if os.path.exists(SEFARIA_DATA_PATH + "/" + format[0]):
+			rmtree(SEFARIA_DATA_PATH + "/" + format[0])
+
+
+def export_links():
+	"""
+	Creates a single CSV file containing all links known to Sefaria.
+	"""
+	with open(SEFARIA_DATA_PATH + "/links/links.csv", 'wb') as csvfile:
+		writer = csv.writer(csvfile)
+		writer.writerow([
+							"Citation 1",
+							"Citation 2",
+							"Type",
+						 ])
+		links = db.links.find().sort([["refs.0", 1]])
+		for link in links:
+			writer.writerow([
+							link["refs"][0],
+							link["refs"][1],
+							link["type"],
+						 ])
 
 
 def export_all():
@@ -91,6 +120,8 @@ def export_all():
 	listed in export_formats.
 	"""
 	clear_exports()
+
+	export_links()
 
 	texts = db.texts.find()
 	for text in texts:
@@ -102,7 +133,7 @@ def export_all():
 		text.update(index)
 		del text["_id"]
 		text["text"] = text.pop("chapter")
-		
+
 		for format in export_formats:
 			out = format[1](text)
 			path = make_path(text, format[0])
